+++ interfaz/admin/barraNavegacionAdmin.php
@@ -8,47 +8,36 @@
     <script src="https://kit.fontawesome.com/7568cd4100.js" crossorigin="anonymous"></script>
 </head>
 <body id="body">
-    <header>
-        <div class="icon_menu">
-            <i class="fas fa-bars" id="btn_open"></i>
-        </div>
-    </header>
     <div class="menu__side" id="menu_side">
         <div class="options__menu">
             <a href="pedidos.php?tipo=prep" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-cubes" title="Pedidos"></i>
-                    <h4>Pedidos</h4>
+                    <i class="fa-solid fa-cubes" title="Pedidos"></i>
                 </div>
             </a>
             <a href="articulos.php" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-box" title="Articulos"></i>
-                    <h4>Articulos</h4>
+                    <i class="fa-solid fa-box" title="Articulos"></i>
                 </div>
             </a>
             <a href="usuarios.php" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-users" title="Usuarios"></i>
-                    <h4>Usuarios</h4>
+                    <i class="fa-solid fa-users" title="Usuarios"></i>
                 </div>
             </a>
             <a href="transporte.php" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-truck-arrow-right" title="Transporte"></i>
-                    <h4>Transporte</h4>
+                    <i class="fa-solid fa-truck-arrow-right" title="Transporte"></i>
                 </div>
             </a>
             <a href="reportes.php" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-book-open" title="Transporte"></i>
-                    <h4>Reportes</h4>
+                    <i class="fa-solid fa-book-open" title="Transporte"></i>
                 </div>
             </a>
             <a href="../../login.php" class="selected">
                 <div class="option">
-                <i class="fa-solid fa-rectangle-xmark" title="Cerrar"></i>
-                    <h4>Cerrar</h4>
+                    <i class="fa-solid fa-rectangle-xmark" title="Cerrar"></i>
                 </div>
             </a>